requirements.txt
asgiref==3.3.1
Django==3.2a1
djangorestframework==3.14.0
pycodestyle==2.11.1
pytz==2021.1
sqlparse==0.4.1
tomli==2.0.1
